Optimizing Angola

+++ catscripts/Government/Countries/Angola/Angola.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i 'Angola' newpages.txt >> Angola.txt
+KEYWORDS_ANGOLA="Angola"
 
-ANGOLA=`stat --print=%s Angola.txt`
-
-if [ $ANGOLA -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Angola.txt"
-  export CATNAME="Angola"
-  $CATEGORIZE
-fi
 
-rm Angola.txt
+  if [ "$DEBUG" ==  "yes" ];
+  then
+    printf "Starting Angola\n"
+  fi
+
+  ANGOLA=`egrep -i "$KEYWORDS_ANGOLA" newpages.txt`
+
+  if [ "$ANGOLA" != "" ];
+  then
+    printf "$ANGOLA" > Angola.txt
+    export CATFILE="Angola.txt"
+    export CATNAME="Angola"
+    $CATEGORIZE
+    rm Angola.txt
+    unset ANGOLA
+  fi
+
+  if [ "$DEBUG" ==  "yes" ];
+  then
+    printf "Finishing Angola\n"
+  fi
+
+fi
